More file renaming
refs: #3203
