+++ translations/geonames_core.js
